requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
tensorflow>=2.20.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

